Moving generated archiver data class to new module
